Do not return paths longer than the output array
GOAP's planner was not checking correctly if the path found was longer
than the provided output array before writing to it. This could create
issue in cases where the planner's search space was large enough to find
a path, but the output array was not big enough to hold it.

GitOrigin-RevId: 2f69402c4939f17b173c9bde063825f9f4eaf083